+++ src/core/haiku/SDL_BApp.h
