+++ src/core/haiku/SDL_BeApp.cc
