+++ src/core/haiku/SDL_BeApp.h
