+++ src/video/haiku/SDL_bvideo.cc
@@ -20,7 +20,7 @@
 */
 
 #include "SDL_internal.h"
-#include "../../main/haiku/SDL_BApp.h"
+#include "../../core/haiku/SDL_BApp.h"
 
 #if SDL_VIDEO_DRIVER_HAIKU
